# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HaloBuzz API
  description: |
    # HaloBuzz Live Streaming & Social Platform API

    The HaloBuzz API provides comprehensive functionality for:
    - **Authentication & User Management** - Secure JWT-based auth with MFA
    - **Live Streaming** - Agora-powered real-time streaming
    - **Social Features** - Follow/unfollow, reels, comments, likes
    - **Gaming Integration** - Multi-game support with real-time multiplayer
    - **Payment System** - Secure transactions, wallets, virtual currency
    - **AI-Powered Features** - Content moderation, recommendations
    - **Real-time Communication** - WebSocket-based chat and notifications

    ## Authentication
    All protected endpoints require JWT token in Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    - Standard endpoints: 100 requests/15 minutes
    - Auth endpoints: 10 requests/15 minutes
    - Payment endpoints: 20 requests/15 minutes

    ## Error Handling
    All errors follow consistent format:
    ```json
    {
      "success": false,
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```

  version: 2.1.0
  termsOfService: 'https://halobuzz.com/terms'
  contact:
    name: HaloBuzz API Support
    url: 'https://halobuzz.com/support'
    email: api-support@halobuzz.com
  license:
    name: Proprietary
    url: 'https://halobuzz.com/license'
servers:
  - url: 'https://api.halobuzz.com/v2'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account with email verification
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              standard_user:
                summary: Standard user registration
                value:
                  username: johndoe
                  email: john@example.com
                  password: SecureP@ss123
                  phone: "+1234567890"
                  country: US
                  language: en
                  agreeToTerms: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
              required: [refreshToken]
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: New access token
                      refreshToken:
                        type: string
                        description: New refresh token (optional)
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - agreeToTerms
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_-]+$'
          description: Unique username (3-30 chars, alphanumeric, underscore, hyphen)
        email:
          type: string
          format: email
          description: Valid email address
        password:
          type: string
          minLength: 8
          description: Password (min 8 chars, must include uppercase, lowercase, number, special char)
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
          description: Optional phone number in E.164 format
        country:
          type: string
          minLength: 2
          maxLength: 2
          description: ISO 3166-1 alpha-2 country code
        language:
          type: string
          default: en
          description: Preferred language (ISO 639-1)
        agreeToTerms:
          type: boolean
          description: Must be true to register
        marketingOptIn:
          type: boolean
          default: false
          description: Opt-in for marketing communications
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or username
        password:
          type: string
          description: User password
        rememberMe:
          type: boolean
          default: false
          description: Extended session duration
        deviceId:
          type: string
          description: Device identifier for security
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
            refreshToken:
              type: string
              description: Refresh token
            expiresIn:
              type: integer
              description: Token expiration in seconds
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
          format: uri
          nullable: true
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          maxLength: 500
          nullable: true
        country:
          type: string
        language:
          type: string
        isVerified:
          type: boolean
        kycStatus:
          type: string
          enum: [pending, verified, rejected]
        ogLevel:
          type: integer
          minimum: 0
          maximum: 100
        coins:
          type: object
          properties:
            balance:
              type: integer
            totalEarned:
              type: integer
            totalSpent:
              type: integer
        trust:
          type: object
          properties:
            score:
              type: integer
              minimum: 0
              maximum: 100
            level:
              type: string
              enum: [novice, trusted, expert, legendary]
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        requestId:
          type: string
          description: Request ID for tracking
  responses:
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              fields:
                - field: "email"
                  message: "Invalid email format"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimit:
      description: Too Many Requests - Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before next request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
            details:
              retryAfter: 900
...